requirements.txt
#pip install -r requirements.txt

pymongo
tabulate
pandas
python-dotenv
flask